renamed queries to AnnouncementQueries
