+++ backend/app.js
@@ -2,7 +2,7 @@ const express = require("express");
 const app = express();
 const bodyParser = require("body-parser");
 const port = 3001;
-const db = require("./queries");
+const db = require("./AnnouncementsQueries");
 const cors = require("cors");
 const readline = require("readline");
 const { google } = require("googleapis");
